Rename tests/gateway to tests/integration/event-service
