+++ tests/integration/event-service/Gopkg.toml
