+++ tests/integration/event-service/entrypoint.sh
